separated textbuttoncontrol from buttoncontrol, added @thhomas fix (missing imports)
